Move PR guidance to interaction owner
