+++ actions/utils/__init__.py
@@ -21,7 +21,6 @@ from .openai_utils import (
     MAX_PROMPT_CHARS,
     filter_labels,
     get_agent_response,
-    get_pr_first_comment_template,
     get_pr_open_response,
     get_pr_summary_guidelines,
     get_pr_summary_prompt,
@@ -52,7 +51,6 @@ __all__ = (
     "format_skipped_files_dropdown",
     "format_skipped_files_note",
     "get_agent_response",
-    "get_pr_first_comment_template",
     "get_pr_open_response",
     "get_pr_summary_guidelines",
     "get_pr_summary_prompt",
